Make output obey selected format for both success and fail responses
Originally, I made fail responses fall back to "reqresp" on the theory it would
allow the user to retry those requests even if they hadn't anticipated dealing
with that by selecting -o reqresp. However, in practice this is just annoying,
and you can't override it even with explicit -o outfmt. A better idiot-proofing
would be to log the reqresp to a /tmp to be recoverable, so consider that for
the future.